refactor: moved EmployeeAvatar and employeeInitials to shared component for better code organization
